# tuma-platform/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Tuma API",
    "description": "Backend API for the Tuma commerce + delivery platform",
    "license": {
      "name": ""
    },
    "version": "0.1.0"
  },
  "paths": {
...
    "/v1/admin/customers": {
      "get": {
        "tags": [
          "admin"
        ],
        "operationId": "list_customers",
        "responses": {
          "200": {
            "description": "All customer accounts, oldest first",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/CustomerAdminResponse"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Not authenticated"
          },
          "403": {
            "description": "Not an admin"
          }
        }
      }
    },
    "/v1/admin/customers/{id}": {
      "delete": {
        "tags": [
          "admin"
        ],
        "summary": "Hard-delete a customer account. The profile, refresh tokens, and order\nhistory follow via ON DELETE CASCADE — permanent, and the platform\nconfirms before calling. Scoped to customer accounts: any other identity\nis a 404.",
        "operationId": "delete_customer",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Customer account id",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Customer deleted"
          },
          "401": {
            "description": "Not authenticated"
          },
          "403": {
            "description": "Not an admin"
          },
          "404": {
            "description": "No customer with that id"
          }
        }
      },
      "patch": {
        "tags": [
          "admin"
        ],
        "operationId": "update_customer",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Customer account id",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateCustomerInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "The updated customer",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CustomerAdminResponse"
                }
              }
            }
          },
          "401": {
            "description": "Not authenticated"
          },
          "403": {
            "description": "Not an admin"
          },
          "404": {
            "description": "No customer with that id"
          },
          "409": {
            "description": "Phone already taken"
          },
          "422": {
            "description": "Invalid input"
          }
        }
      }
    },
    "/v1/admin/merchants": {
      "get": {
        "tags": [
          "admin"
        ],
        "operationId": "list_merchants",
        "responses": {
          "200": {
            "description": "All merchant businesses, oldest first",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/MerchantResponse"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Not authenticated"
          },
          "403": {
            "description": "Not an admin"
          }
        }
      },
      "post": {
        "tags": [
          "admin"
        ],
        "operationId": "create_merchant",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateMerchantInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Merchant business created with its owner account",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreatedMerchantResponse"
                }
              }
            }
          },
          "401": {
            "description": "Not authenticated"
          },
          "403": {
            "description": "Not an admin"
          },
          "409": {
            "description": "Owner email already taken"
          },
          "422": {
            "description": "Invalid input"
          }
        }
      }
    },
    "/v1/admin/merchants/{id}": {
      "get": {
        "tags": [
          "admin"
        ],
        "summary": "One business with everything the admin detail page shows: the business,\nits stores with assortment counts, and its members. Admin sees facts;\nthe merchant manages the menu — it stays out of the admin's response.",
        "operationId": "get_merchant",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Merchant business id",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The business, its stores, and its members",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MerchantDetailResponse"
                }
              }
            }
          },
          "401": {
            "description": "Not authenticated"
          },
          "403": {
            "description": "Not an admin"
          },
          "404": {
            "description": "No merchant business with that id"
          }
        }
      },
      "delete": {
        "tags": [
          "admin"
        ],
        "summary": "Hard-delete a business. Its memberships, stores (and their assortments),\nand catalog products follow via ON DELETE CASCADE. Member ACCOUNTS are\nidentities, not parts of the business — they survive the delete.",
        "operationId": "delete_merchant",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Merchant business id",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Merchant business deleted"
          },
          "401": {
            "description": "Not authenticated"
          },
          "403": {
            "description": "Not an admin"
          },
          "404": {
            "description": "No merchant business with that id"
          }
        }
      },
      "patch": {
        "tags": [
          "admin"
        ],
        "operationId": "update_merchant",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Merchant business id",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateMerchantInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "The updated business",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MerchantResponse"
                }
              }
            }
          },
          "401": {
            "description": "Not authenticated"
          },
          "403": {
            "description": "Not an admin"
          },
          "404": {
            "description": "No merchant business with that id"
          },
          "422": {
            "description": "Invalid input"
          }
        }
      }
    },
    "/v1/admin/payments/{id}/collect": {
      "post": {
        "tags": [
          "admin"
        ],
        "summary": "Mark a pending cash payment collected. Its allocations settle with it —\nthe ledger records who received what without any inference.",
        "operationId": "collect_payment",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Payment id",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Payment collected",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaymentResponse"
                }
              }
            }
          },
          "400": {
            "description": "Only a pending payment can be collected"
          },
          "401": {
            "description": "Not authenticated"
          },
          "403": {
            "description": "Not an admin"
          },
          "404": {
            "description": "No payment with that id"
          }
        }
      }
    },
    "/v1/admin/store-orders/{id}": {
      "patch": {
        "tags": [
          "admin"
        ],
        "summary": "Advance any store order. The platform's control plane answer when a\nmerchant cannot.",
        "operationId": "advance_store_order_admin",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Store order id",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AdvanceStatusInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Store order advanced",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MerchantStoreOrderResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid or illegal status transition"
          },
          "401": {
            "description": "Not authenticated"
          },
          "403": {
            "description": "Not an admin"
          },
          "404": {
            "description": "No store order with that id"
          }
        }
      }
    },
    "/v1/admin/summary": {
      "get": {
        "tags": [
          "admin"
        ],
        "summary": "Platform-wide counts for the admin dashboard.",
        "operationId": "summary",
        "responses": {
          "200": {
            "description": "Platform-wide counts",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminSummaryResponse"
                }
              }
            }
          },
          "401": {
            "description": "Not authenticated"
          },
          "403": {
            "description": "Not an admin"
          }
        }
      }
    },
    "/v1/auth/login": {
      "post": {
        "tags": [
          "auth"
        ],
        "summary": "Platform sign-in (merchant operators and admins) with email + password.\nSets the two session cookies (15-min access JWT + 30-day opaque refresh);\nthe body is empty. Authorization is resolved from profiles — an account\nwith neither an admin profile nor a merchant membership cannot enter the\nplatform, and the answer never reveals which part failed.",
        "operationId": "login",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": "Session cookies set (tuma-auth-token + tuma-auth-refresh)"
          },
          "401": {
            "description": "One generic error for every failure path"
          },
          "422": {
            "description": "Invalid input"
          }
        }
      }
    },
    "/v1/auth/logout": {
      "post": {
        "tags": [
          "auth"
        ],
        "operationId": "logout",
        "responses": {
          "204": {
            "description": "Logged out"
          },
          "401": {
            "description": "Not authenticated"
          }
        }
      }
    },
    "/v1/auth/otp/request": {
      "post": {
        "tags": [
          "auth"
        ],
        "operationId": "otp_request",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OtpRequestInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Always the same response — reveals nothing about the phone number",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OtpRequestResponse"
                }
              }
            }
          },
          "422": {
            "description": "Invalid phone number"
          }
        }
      }
    },
    "/v1/auth/otp/verify": {
      "post": {
        "tags": [
          "auth"
        ],
        "operationId": "otp_verify",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OtpVerifyInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Code valid — account and customer profile fetched or created, token issued",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OtpVerifyResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid or expired code"
          },
          "422": {
            "description": "Invalid input"
          }
        }
      }
    },
    "/v1/auth/password": {
      "post": {
        "tags": [
          "auth"
        ],
        "summary": "Change the caller's own password. Merchant operators and admins only in\npractice — customers have no password (OTP accounts), which is a 400.",
        "operationId": "change_password",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChangePasswordInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": "Password changed"
          },
          "400": {
            "description": "Account has no password (OTP sign-in)"
          },
          "401": {
            "description": "Not authenticated, or current password is wrong"
          },
          "422": {
            "description": "Invalid input"
          }
        }
      }
    },
    "/v1/me": {
      "get": {
        "tags": [
          "auth"
        ],
        "operationId": "me",
        "responses": {
          "200": {
            "description": "The authenticated account with its profiles and memberships",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MeResponse"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid credentials"
          }
        }
      },
      "patch": {
        "tags": [
          "auth"
        ],
        "summary": "Updates the caller's own profile name — the customer profile when there\nis one, else the admin profile. Merchant-operator accounts have no\neditable name in V1 (their identity on the wing is the business).",
        "operationId": "update_me",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateMeInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "The updated profile",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MeResponse"
                }
              }
            }
          },
          "400": {
            "description": "Empty name, or no editable profile"
          },
          "401": {
            "description": "Missing or invalid credentials"
          },
          "422": {
            "description": "Invalid input"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AdvanceStatusInput": {
        "type": "object",
        "required": [
          "status"
        ],
        "properties": {
          "status": {
            "type": "string"
          }
        }
      },
...
      "CustomerAdminResponse": {
        "type": "object",
        "description": "A customer account on the admin's list: the profile (name) joined with\nthe account facts (phone, active flag).",
        "required": [
          "user_id",
          "is_active",
          "created_at"
        ],
        "properties": {
          "created_at": {
            "type": "string",
            "format": "date-time"
          },
          "is_active": {
            "type": "boolean"
          },
          "name": {
            "type": [
              "string",
              "null"
            ]
          },
          "phone": {
            "type": [
              "string",
              "null"
            ]
          },
          "user_id": {
            "type": "string",
            "format": "uuid"
          }
        }
      },
...
      "UpdateCustomerInput": {
        "type": "object",
        "description": "Admin edits to a customer account. Provided fields overwrite, an empty\nname clears it, absent fields keep their value. Phone stays editable\nbecause a typo'd number during OTP signup is exactly the kind of thing\nan admin corrects; a taken phone is the typed 409.",
        "properties": {
          "is_active": {
            "type": [
              "boolean",
              "null"
            ]
          },
          "name": {
            "type": [
              "string",
              "null"
            ]
          },
          "phone": {
            "type": [
              "string",
              "null"
            ]
          }
        }
      },
...
    }
  }
}